修改命令逻辑

+++ src/Console/Commands/Route.php
@@ -12,28 +12,49 @@ class Route extends Command
 
     public function out()
     {
-        echo $this->out;
-    }
-
-    public function l()
-    {
-        \Yao\Facade\Route::register();
-        return print_r(array_filter(\Yao\Facade\Route::getRoute()), true);
-    }
-
-    public function cache()
-    {
-        \Yao\Facade\Route::register();
-        file_put_contents(self::ROUTEFILE, serialize(array_filter(\Yao\Facade\Route::getRoute())));
-        return "缓存生成成功\n";
-    }
-
-    public function dcache()
-    {
-        if (!file_exists(self::ROUTEFILE)) {
-            return "没有缓存文件！\n";
+        echo <<<EOT
+        (1). 输出路由列表
+        (2). 生成路由缓存
+        (3). 删除路由缓存
+        (4). 退出
+请输入选项[1,2,3,4]:
+EOT;
+        while (1) {
+            fscanf(STDIN, '%d', $options);
+            if (!empty($options)) {
+                switch ($options) {
+                    case 1:
+                        \Yao\Facade\Route::register();
+                        echo "——————————————————————————————————————————————————————————————————————————————————\n";
+                        foreach (\Yao\Facade\Route::getRoute() as $method => $routes) {
+                            foreach ($routes as $route => $locate) {
+                                if (is_array($locate['route'])) {
+                                    $locate['route'] = implode('->', $locate['route']);
+                                }
+                                echo str_pad(strtoupper($method), 5, ' ', STR_PAD_BOTH) . '|' . str_pad($route, 20, ' ', STR_PAD_BOTH) . '|' . str_pad($locate['route'], 54, ' ', STR_PAD_BOTH) . "\n";
+                                echo "——————————————————————————————————————————————————————————————————————————————————\n";
+                            }
+                        }
+                        exit;
+                    case 2:
+                        \Yao\Facade\Route::register();
+                        if(!file_exists(dirname(self::ROUTEFILE))){
+                            mkdir(dirname(self::ROUTEFILE),0777,true);
+                        }
+                        file_put_contents(self::ROUTEFILE, serialize(array_filter(\Yao\Facade\Route::getRoute())));
+                        exit("缓存生成成功\n");
+                    case 3:
+                        if (!file_exists(self::ROUTEFILE)) {
+                            exit("没有缓存文件！\n");
+                        }
+                        unlink(self::ROUTEFILE);
+                        exit("缓存生成已经清除\n");
+                    case 4:
+                        exit;
+                    default:
+                        echo "输入错误！\n重新输入选项[输入4退出]：";
+                }
+            }
         }
-        unlink(self::ROUTEFILE);
-        return "缓存生成已经清除\n";
     }
 }
